+++ .devcontainer/devcontainer.json
@@ -1,5 +1,5 @@
 {
-    "name": "python_template",
+    "name": "mcp_devtasks",
     "build": {
         "dockerfile": "Dockerfile",
         "context": "..",
