+++ .vscode/settings.json
@@ -8,7 +8,7 @@
 	},
 	"files.exclude": {
 		"**/__pycache__/**": true,
-		"/home/vscode/.local/lib/python3.10/site-packages/python_template/**": true
+		"/home/vscode/.local/lib/python3.10/site-packages/mcp_devtasks/**": true
 	},
 	"python.analysis.autoImportCompletions": false //vscode gets it wrong more than right and mostly gets in the way
 }
