+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
 
-## python_template
+## mcp_devtasks
 
 ## [0.0.0]
